backend/requirements.txt
# Core FastAPI dependencies
fastapi>=0.100.0
uvicorn[standard]>=0.20.0

# LangChain and AI dependencies
langchain>=0.0.300
langchain-core>=0.0.10
langchain-google-genai>=0.0.5

# Data validation and environment
pydantic>=2.0.0
python-dotenv>=1.0.0

# Alternative: Use these versions if the above fail
# fastapi==0.104.1
# uvicorn[standard]==0.24.0
# langchain==0.1.0
# langchain-core==0.1.0
# langchain-google-genai==0.0.5
# pydantic==2.5.0
# python-dotenv==1.0.0
